refactor: Change BoardScrap Dir (#67)
