+++ src/main/java/paratrip/paratrip/scrap/board/controller/vo/response/BoardScrapResponseVo.java
@@ -1,4 +1,4 @@
-package paratrip.paratrip.scrap.controller.vo.response;
+package paratrip.paratrip.scrap.board.controller.vo.response;
 
 public class BoardScrapResponseVo {
 	public record AddBoardScrapResponse(
