+++ src/main/java/paratrip/paratrip/scrap/board/mapper/BoardScrapMapper.java
@@ -1,4 +1,4 @@
-package paratrip.paratrip.scrap.mapper;
+package paratrip.paratrip.scrap.board.mapper;
 
 import org.mapstruct.InjectionStrategy;
 import org.mapstruct.Mapper;
@@ -6,7 +6,7 @@ import org.mapstruct.Mapping;
 import org.mapstruct.ReportingPolicy;
 
 import paratrip.paratrip.board.main.entity.BoardEntity;
-import paratrip.paratrip.scrap.entity.BoardScrapEntity;
+import paratrip.paratrip.scrap.board.entity.BoardScrapEntity;
 import paratrip.paratrip.member.entity.MemberEntity;
 
 @Mapper(componentModel = "spring", injectionStrategy = InjectionStrategy.CONSTRUCTOR, unmappedTargetPolicy = ReportingPolicy.IGNORE)
